product update

+++ routes/api/product.routes.js
@@ -47,29 +47,13 @@ router.post(
 			case 'createProduct':
 				handleRequest(req, res, async () => {
 					console.log('attempting to create product')
-					const {
-						visibility,
-						productId,
-						name,
-						description,
-						price,
-						attributes,
-						category,
-						imageUrl,
-					} = req.body.product
 					const authenticatedUserId = req.user.uid
+					const product = req.body.product
 					validateRequestBody(req.body, ['product']) // Required fields for creating a product
 
 					const result = await ProductDB.createProduct(
 						authenticatedUserId,
-						visibility,
-						productId,
-						name,
-						description,
-						price,
-						attributes,
-						category,
-						imageUrl
+						product
 					)
 
 					return result
@@ -91,29 +75,13 @@ router.put(
 			case 'updateProduct':
 				handleRequest(req, res, async () => {
 					console.log('attempting to update product')
-					const {
-						visibility,
-						name,
-						productId,
-						description,
-						price,
-						attributes,
-						category,
-						imageUrl,
-					} = req.body.product
 					const authenticatedUserId = req.user.uid
+					const product = req.body.product
 					validateRequestBody(req.body, ['product']) // Required fields for creating a product
 
 					const result = await ProductDB.updateProduct(
 						authenticatedUserId,
-						productId,
-						visibility,
-						name,
-						description,
-						price,
-						attributes,
-						category,
-						imageUrl
+						product
 					)
 
 					return result
@@ -132,23 +100,22 @@ router.delete(
 		const endpoint = decodeEndpoint(req)
 
 		switch (endpoint) {
-			case 'deleteProduct':
-				handleRequest(req, res, async () => {
-					console.log('attempting to delete a product')
-					const { productId } = req.params
-					const authenticatedUserId = req.user.uid
-					validateParams(req.params, ['productId'])
-
-					const result = await ProductDB.deleteProduct(
-						authenticatedUserId,
-						productId
-					)
-
-					return result
-				})
-				break
 			default:
-				return handleResponse({ message: 'Invalid endpoint' }, true)
+				if (endpoint.startsWith('deleteProduct/')) {
+					handleRequest(req, res, async () => {
+						console.log('attempting to delete a product by id')
+						const authenticatedUserId = req.user.uid
+						const productId = endpoint.split('/')[1]
+						validateParams({ productId }, ['productId'])
+
+						return await ProductDB.deleteProduct(
+							authenticatedUserId,
+							productId
+						)
+					})
+				} else {
+					return handleResponse({ message: 'Invalid endpoint' }, true)
+				}
 		}
 	}
 )
